add FrameDisassembler

+++ src/constants.zig
@@ -7,4 +7,4 @@ pub const frame_headers_magic: u16 = 0x4B42;
 pub const chunk_data_size: usize = 1024 * 4;
 
 /// Maximum size of a frame's payload
-pub const max_frame_payload_size: usize = std.math.maxInt(u16);
+pub const max_frame_payload_size: u16 = std.math.maxInt(u16);
